messageTray: Allow empty notification sources
Currently, a notification source is automatically destroyed once its
last notification is destroyed. That makes sense for most notification
sources.

However, more complex notification sources might have internal state
tracking and other calculations going on, about deciding when to next
show a notification. It’s useful to not tear down that state, so add a
property to `NotificationSource` to prevent it being destroyed when
empty, if desired.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>