messageTray: Remove expand on hover and simplify how notifications are dismissed
This simplifies the code of tracking hover and timeouts a lot, by
removing the expand on hover and the special tracking of the hover
state. We now have an expand button instead to allow users to expand the
notification, which also fixed touch.
This also reduces the timeout before a banner hides because now that we
allow expanding in the notification drawer it's not as important anymore
that the user clicks on buttons in the banner and they can have the same
interaction later as well.