layout: Add CSS classes to adapt style based on the monitor size
We'll need to adapt the spacing (and potentially the icon sizes) of the icon
grid based on the monitor size, so that the grid fits better on small
monitors. The easy way to do this (and also the way everyone does it these
days) is to add CSS classes based on the monitor size to the uiGroup, aka a
very simple version of layout breakpoints. These classes can simply be
selected via the selector "#uiGroup.small-screen-size" in CSS, and therefore
styling can be made dependent on the screen size.

The small/normal/large screen classes are chosen based on the dimension with
the least available space for now. That might not be the most flexible
approach, but seems simple enough right now. Also the threshold sizes are
not set in stone and can be tweaked again later of course.