style/app-grid: Adjust appGrid sizing based on "breakpoints" for screen size
We just introduced a simple version of "breakpoints" to be able to style
things differently in CSS based on the screen size. Let's start using it and
change the various spacings in the appGrid dependent on the size of the
monitor.

For small screens, also use a smaller font size, so that we're able to
display slightly longer (about 3 more characters) app titles on those
small monitors.